# dedup: use structural sha256 hash instead of agnostify and serialize