… benchmark/src/main/scala/firrtl/benchmark/hot/TransformBenchmark.scala
… src/main/scala/firrtl/transforms/Dedup.scala
… src/test/scala/firrtl/ir/StructuralHashSpec.scala