@ekiwi-sifive 2020-06-30
I still need to actually measure the performance improvement that I am hoping for.