@seldridge 2020-06-30
This looks cool. One nit: maybe SHA-256 is a better choice than MD5? MD5 (or SHA-1) seems to beg for the DAC 2021 submission: "Reusability is FIRRTL Ground *for Hardware Trojans*".

Does choice of different hashing algorithms markedly affect the benchmarking results?