@ekiwi-sifive 2020-06-30
> Does choice of different hashing algorithms markedly affect the benchmarking results?

I can try that out. However, note that the previous hash used was the standard `String.hashCode`, so MD5 is already a huge improvement. Jack and I were even thinking to switch to a 128-bit MurMur3 hash (which would require a new dependency for firrtl).