@ekiwi-sifive 2020-06-30
Initial Benchmarking Results on my Mac:


  | old (ms) | new (ms) | fraction of old time
-- | -- | -- | --
small @ JDK1.8 | 2523 | 1085 | 43%
small @ JDK14 | 2589 | 978 | 38%
small @ Graal | 2271 | 853 | 38%
medium @ JDK1.8 | 11596 | 8010 | 69%
medium @ JDK14 | 10639 | 7212 | 68%
medium @ Graal | 9861 | 5744 | 58%