@ekiwi-sifive 2020-06-30
Looking at a flame graph for the small design: in the Dedup pass only ~20% of the time is now spent in `buildRTLTags` which is the method improved by this PR.