@ekiwi-sifive 2020-06-30
>  ~78% of the time in Dedup is spent deduping the annotations for the `medium` design

In case anyone is interested in further improving Dedup performance: Most (like 99%) of these 78% is spent in `InstanceGraph.findInstancesInHierarchy`.