@ekiwi-sifive 2020-07-01
> (1) Have you considered making the MD5 generic?

This hash turned out to be pretty specialized, i.e., it only really makes sense to run it on `DefModule` nodes which is why I ended up making the method that operates on arbitrary `FirttlNode`s package private. The way we agnostify references makes sense in the context of hashing a module but might have pretty unintuitive behavior when hashing expressions directly.

> (2) Looking at the ScalaDoc for `MurmurHash3`, it links to https://github.com/aappleby/smhasher which has a test suite for the distribution, collision, and performance of non-cryptographic hash functions. Probably a bit hard to wield here, but may be worth considering trying to measure some of these properties to check our implementation.

I am not sure how this would apply. The underlying hash function that I am using is md5 which has known short comings but should be good enough for this application where we do not expect malicious input. There might be bugs where we hash two nodes to the same value even though be consider them structurally non-equivalent (e.g. if there is a type and an `id` is reused for different node types). However, I am not sure how the linked test suite would find bugs like that.

I am going to look at addressing your comment on the code.