@jackkoenig 2020-07-01
> I am not sure how this would apply. The underlying hash function that I am using is md5 which has known short comings but should be good enough for this application where we do not expect malicious input. 

I think it applies in checking that the implementation is correct; if we get it wrong and there are tons of collisions it could be a (very) hard-to-diagnose performance issue and even a correctness issue if we rely on these hashes for equality checks.