@ekiwi-sifive 2020-07-02
@jackkoenig I changed the code to ignore block structure and empty statements as we discussed.