@ekiwi-sifive 2020-07-02
@jackkoenig or @seldridge : maybe you can answer this question I had while developing this:
```
// TODO: should extmodule portnames always be significant since they map to the verilog pins?
```

I.e., should you be able to dedup external modules that have the same defname but different port names? I guess this would be a bug in any case, so no sensible circuit should be affected by this question, but what do you think?