@seldridge 2020-07-02
External modules with the same defname and different ports smells like something that `CheckHighForm` should catch. I think it's reasonable to assume that that isn't possible to exist when dedup runs.