@ekiwi-sifive 2020-07-02
@seldridge the last two commits should address your comments.