@ekiwi-sifive 2020-07-08
Thanks for your review @seldridge ! Please let me know what to do about the `DebugHasher` then I can implement all the fixes.