@ekiwi-sifive 2020-07-08
I addressed @seldridge 's comments in the latest commit.